backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
openai
aiosqlite
python-multipart
fpdf2
httpx
